// app/containers/View/types.ts
export interface IViewBase {
  id: number
  name: string
  description: string
  sourceName: string
  [propName: string]: any
}

type IViewTemp = Omit<IViewBase, 'sourceName'>

export interface IView extends IViewTemp {
  sql: string,
  sourceSql: string,
  model: string
  variable: string
  config: string
  projectId: number
  source?: ISourceSimple
  sourceId: number
  roles: IViewRoleRaw[]
  rename?: string
  cronExpression?:string
}
// ...
export interface ISqlColumn {
  name: string
  type: SqlTypes
}

export interface IExecuteSqlResponse {
  columns: ISqlColumn[]
  totalCount: number
  pageNo: number
  resultList: Array<{ [key: string]: string | number }>
}

export interface IExcuteSqlStatisticsResponse<T> {
  data: T,
  status: number
}


export interface IViewModelProps {
  name: string
  sqlType: SqlTypes
  visualType: ViewModelVisualTypes
  modelType: ViewModelTypes
}

export type IKeyOfViewModelProps = keyof Omit<IViewModelProps, 'name'>

export interface IViewModel {
  [name: string]: Omit<IViewModelProps, 'name'>
}

interface IViewVariableChannel {
  bizId: number
  name: string
  tenantId: number
}

interface IViewVariableBase {
  name: string
  type: ViewVariableTypes
  valueType: ViewVariableValueTypes
  defaultValues: Array<string | number | boolean>
  channel?: IViewVariableChannel
  udf: boolean
}

export interface IViewVariable extends IViewVariableBase {
  key: string
  alias: string
  fromService: boolean
}

export interface IViewRoleRaw {
  roleId: number
  columnAuth: string
  rowAuth: string
}

export interface IViewRoleRowAuth {
  name: string
  values: Array<string | number | boolean>
  enable: boolean
}

export interface IViewRole {
  roleId: number
  /**
   * view columns name
   * @type {string[]}
   * @memberof IViewRole
   */
  columnAuth: string[]

  /**
   * query variable values
   * @type {(Array<string | number>)}
   * @memberof IViewRole
   */
  rowAuth: IViewRoleRowAuth[]
}